requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3